Fixes
Fixed some BigDecimal Math to now allow crazy decimals (like percent / 100.0)
Fixed RentalDay Calculation Loop to properly go from day AFTER checkout to due
Added some basic abstraction to checkout/rentalcheckout

+++ lib/src/main/java/bw1021/exceptions/InvalidDiscountException.java
@@ -1,16 +1,14 @@
 package bw1021.exceptions;
 
-import java.math.BigDecimal;
-
 public class InvalidDiscountException extends Exception {
 	
 	/**
 	 * 
 	 */
 	private static final long serialVersionUID = 1746510738834168055L;
-	private BigDecimal discountPercent;
+	private int discountPercent;
 	
-	public InvalidDiscountException(BigDecimal discountPercent) {
+	public InvalidDiscountException(int discountPercent) {
 		this.discountPercent = discountPercent;
 		
 	}
